+++ RELEASENOTES.md
@@ -4,6 +4,7 @@ The release notes are summaries of important changes and fixes in each release.
 There is maximum one release per date, but this can be performed as one or several deployments during a day.
 ## Release 1.9.1
 Release date 2023-08-31
+- **Icons** for meeting status with explaning tootip.
 - **Bug fix** of module ownership transfer and assignment of assistant maintainer.
 ## Release 1.9.0
 Release date 2023-08-30 - Milestone: Fully support for wiFRED management.